preventing GIL contention, by .sleep(0) inside flow_euler.py at every iter

+++ trellis2/modules/utils.py
@@ -72,7 +72,7 @@ def manual_cast(tensor, dtype):
     Cast if autocast is not enabled.
     """
     if not torch.is_autocast_enabled():
-        return tensor.type(dtype)
+        return tensor.to(dtype)
     return tensor
